Add $match operation and test.
